refactor: moved code from /cytoscape to root page /
